added `NODES_TIMING` option to log execution time of each node
